+++ src/index.ts
@@ -5,7 +5,7 @@ import invocationValidator from './invocationValidator';
 import { accountStep } from './steps/account';
 import { containerSteps } from './steps/containers';
 import { scanSteps } from './steps/scans';
-import { userStep } from './steps/users';
+import { userStep } from './steps/access';
 
 export const invocationConfig: IntegrationInvocationConfig<TenableIntegrationConfig> =
   {
